server/requirements.txt
python-dateutil==1.5
flask==0.9